+++ tests/Unit/Repository/Member/Field/FieldFactoryTest.php
@@ -1,17 +1,10 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: cyrillbolliger
- * Date: 27.10.18
- * Time: 17:02
- */
 
 namespace App\Repository\Member\Field;
 
 use App\Exceptions\MemberUnknownFieldException;
 use App\Exceptions\MultiSelectOverwriteException;
-use App\Exceptions\WeblingFieldMappingConfigException;
-use Illuminate\Support\Facades\Config;
+use App\Repository\Member\Field\Mapping\Loader;
 use Tests\TestCase;
 
 class FieldFactoryTest extends TestCase {
@@ -26,86 +19,7 @@ class FieldFactoryTest extends TestCase {
 	const TEXT_FIELD = 'lastName';
 	const SKIP_FIELD = 'dontUse';
 	
-	public function test__constructConfigNotFound() {
-		// remove existing instances, else getInstance will return the existing instance
-		/** @noinspection PhpUnhandledExceptionInspection */
-		$instanceField = $this->getPrivateProperty( FieldFactory::class, 'instance' );
-		$instanceField->setValue( null );
-		
-		Config::set( 'app.webling_field_mappings_config_path', 'unknown' );
-		
-		$this->expectException( WeblingFieldMappingConfigException::class );
-		$this->expectExceptionMessage( 'The Webling field mappings config file was not found.' );
-		/** @noinspection PhpUnhandledExceptionInspection */
-		FieldFactory::getInstance();
-	}
-	
-	public function test__constructParseException() {
-		// remove existing instances, else getInstance will return the existing instance
-		/** @noinspection PhpUnhandledExceptionInspection */
-		$instanceField = $this->getPrivateProperty( FieldFactory::class, 'instance' );
-		$instanceField->setValue( null );
-		
-		Config::set( 'app.webling_field_mappings_config_path',
-			$this->getFileRelPath() . DIRECTORY_SEPARATOR . 'webling-field-mappings-parse-error.yml' );
-		
-		$this->expectException( WeblingFieldMappingConfigException::class );
-		$this->expectExceptionMessageRegExp( "/^YAML parse error:/" );
-		/** @noinspection PhpUnhandledExceptionInspection */
-		FieldFactory::getInstance();
-	}
-	
-	private function getFileRelPath() {
-		return str_replace( base_path() . '/', '', dirname( __FILE__ ) );
-	}
-	
-	public function test__constructMappingsException() {
-		// remove existing instances, else getInstance will return the existing instance
-		/** @noinspection PhpUnhandledExceptionInspection */
-		$instanceField = $this->getPrivateProperty( FieldFactory::class, 'instance' );
-		$instanceField->setValue( null );
-		
-		Config::set( 'app.webling_field_mappings_config_path',
-			$this->getFileRelPath() . DIRECTORY_SEPARATOR . 'webling-field-mappings-mappings-not-found.yml' );
-		
-		$this->expectException( WeblingFieldMappingConfigException::class );
-		$this->expectExceptionMessage( 'The entry point ("mappings") was not found or empty.' );
-		/** @noinspection PhpUnhandledExceptionInspection */
-		FieldFactory::getInstance();
-	}
-	
-	public function test__constructInvalidConfigException() {
-		// remove existing instances, else getInstance will return the existing instance
-		/** @noinspection PhpUnhandledExceptionInspection */
-		$instanceField = $this->getPrivateProperty( FieldFactory::class, 'instance' );
-		$instanceField->setValue( null );
-		
-		Config::set( 'app.webling_field_mappings_config_path',
-			$this->getFileRelPath() . DIRECTORY_SEPARATOR . 'webling-field-mappings-invalid-config.yml' );
-		
-		$this->expectException( WeblingFieldMappingConfigException::class );
-		$this->expectExceptionMessageRegExp( "/^Invalid Webling field mapping config:/" );
-		/** @noinspection PhpUnhandledExceptionInspection */
-		FieldFactory::getInstance();
-	}
-	
-	public function test__constructReservedFieldNameException() {
-		// remove existing instances, else getInstance will return the existing instance
-		/** @noinspection PhpUnhandledExceptionInspection */
-		$instanceField = $this->getPrivateProperty( FieldFactory::class, 'instance' );
-		$instanceField->setValue( null );
-		
-		Config::set( 'app.webling_field_mappings_config_path',
-			$this->getFileRelPath() . DIRECTORY_SEPARATOR . 'webling-field-mappings-reserved-field-key.yml' );
-		
-		$this->expectException( WeblingFieldMappingConfigException::class );
-		$this->expectExceptionMessageRegExp( "/^Reserved field key:/" );
-		/** @noinspection PhpUnhandledExceptionInspection */
-		FieldFactory::getInstance();
-	}
-	
 	public function testCreateByInternalKey() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::INTERNAL_FIELD_NAME );
@@ -113,7 +27,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateByWeblingKey() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::WEBLING_FIELD_NAME );
@@ -121,7 +34,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateMemberUnknownFieldException() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		$this->expectException( MemberUnknownFieldException::class );
 		/** @noinspection PhpUnhandledExceptionInspection */
@@ -129,7 +41,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateWithValue() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::WEBLING_FIELD_NAME, 'Hans Muster' );
@@ -137,7 +48,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateDateField() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::DATE_FIELD );
@@ -145,7 +55,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateLongTextField() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::LONG_TEXT_FIELD );
@@ -153,7 +62,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateMultiSelectField() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::MULTI_SELECT_FIELD );
@@ -161,7 +69,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateMultiSelectFieldOverwriteException() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		$this->expectException( MultiSelectOverwriteException::class );
 		/** @noinspection PhpUnhandledExceptionInspection */
@@ -169,7 +76,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateSelectField() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::SELECT_FIELD, self::SELECT_FIELD_VALUE_KEY );
@@ -180,7 +86,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateTextField() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::TEXT_FIELD );
@@ -188,7 +93,6 @@ class FieldFactoryTest extends TestCase {
 	}
 	
 	public function testCreateSkipField() {
-		/** @noinspection PhpUnhandledExceptionInspection */
 		$fieldFactory = FieldFactory::getInstance();
 		/** @noinspection PhpUnhandledExceptionInspection */
 		$field = $fieldFactory->create( self::SKIP_FIELD );
@@ -197,10 +101,10 @@ class FieldFactoryTest extends TestCase {
 	
 	public function testAllConfigMappings() {
 		/** @noinspection PhpUnhandledExceptionInspection */
+		$loader    = Loader::getInstance();
+		$fieldKeys = $loader->getFieldKeys();
+		
 		$fieldFactory = FieldFactory::getInstance();
-		/** @noinspection PhpUnhandledExceptionInspection */
-		$mappings  = $this->getPrivateProperty( FieldFactory::class, 'mappings' );
-		$fieldKeys = array_keys( $mappings->getValue( $fieldFactory ) );
 		
 		foreach ( $fieldKeys as $key ) {
 			/** @noinspection PhpUnhandledExceptionInspection */
@@ -210,25 +114,5 @@ class FieldFactoryTest extends TestCase {
 				$this->assertTrue( $field instanceof Field );
 			}
 		}
-	}
-	
-	/**
-	 * getPrivateProperty
-	 *
-	 * @author    Joe Sexton <joe@webtipblog.com>
-	 *
-	 * @param    string $className
-	 * @param    string $propertyName
-	 *
-	 * @return    \ReflectionProperty
-	 * @throws \ReflectionException
-	 */
-	public function getPrivateProperty( $className, $propertyName ) {
-		/** @noinspection PhpUnhandledExceptionInspection */
-		$reflector = new \ReflectionClass( $className );
-		$property  = $reflector->getProperty( $propertyName );
-		$property->setAccessible( true );
-		
-		return $property;
 	}
 }
